Refactoring: creating separate Component (AddTask) for creating new task. Making Modal Component independent
