+++ index.html
@@ -58,6 +58,7 @@
     <link href="css/Input.css" rel="stylesheet" type="text/css">
     <link href="css/Button.css" rel="stylesheet" type="text/css">
     <link href="css/Modal.css" rel="stylesheet" type="text/css">
+    <link href="css/AddTask.css" rel="stylesheet" type="text/css">
 
     <title>Quantori Homework #5</title>
 </head>
@@ -73,6 +74,7 @@
 <script src="functional/components/App.js" type="module"></script>
 
 <script src="functional/components/Input.js" type="module"></script>
+<script src="functional/components/AddTask.js" type="module"></script>
 <script src="functional/components/TaskTag.js" type="module"></script>
 <script src="functional/components/Modal.js" type="module"></script>
 <script src="functional/components/TaskList.js" type="module"></script>
